# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "SFTAS Zamfara API",
    "description": "Brainstorm.ng Documentation for SFTAS Zamfara API",
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "info@brainstorm.ng",
      "email": "info@brainstorm.ng",
      "url": "https://brainstorm.ng.com"
    }
  },
  "paths": {
...
    "/api/v1/user/verify": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Verify User Token",
        "description": "Endpoint to verify user authentication token",
        "operationId": "verifyUserToken",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "Authentication Key",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "Invalid token"
          }
        }
      }
    },
    "/api/v1/users": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "List of users",
        "description": "List of registed users",
        "operationId": "verifyUserToken",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "Authentication Key",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "User": {
      "required": [
        "name",
        "email",
        "phone",
        "position",
        "accessTo",
        "role",
        "department",
        "accessToDept",
        "password"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "uniqueItems": true
        },
        "phone": {
          "type": "string"
        },
        "position": {
          "type": "string"
        },
        "accessTo": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "department": {
          "type": "string"
        },
        "accessToDept": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "Users": {
      "type": "array",
      "$ref": "#/definitions/User"
    },
...
  },
  "host": "yge.wvi.mybluehost.me/sftas/zamfara",
  "basePath": "",
  "schemes": [
    "https"
  ]
}